Complete transition to dynamic loading solution
All logic with large dependencies is now in .so files (.dylib on mac).
Unit tests verify that this basically works but further work on the
Swift runtime is needed to run integration tests with it.  Even then,
the tests will use docker locally.  It can't be tested in an actual
Nimbella stack without the ability to either deploy a Nimbella test
project or install new runtimes in an existing one.